+++ CHANGELOG.md
@@ -13,7 +13,7 @@ OpenAnalytics v1.0.01 - 02 February 2014
 
  - Initial release with required functionality added
  - Added "/analytics admin" command for setting config parameters in game
- - Added "/anaytics optout" and "/analytics optin" to allow players to opt out/in
+ - Added "/analytics optout" and "/analytics optin" to allow players to opt out/in
  - Added permissions relating to these commands
  - Refactored all common code out to NyvariaComponents library
  - Added support for SignShop (not sure if Google Analytics is picking it up correctly)